save selected travel info to local storage from the api
